Clean unused code
Clean unused imports and unused type conversions.

+++ src/main/java/io/github/mirromutth/r2dbc/mysql/codec/ByteCodec.java
@@ -93,7 +93,7 @@ final class ByteCodec extends AbstractPrimitiveCodec<Byte> {
 
         @Override
         public int hashCode() {
-            return (int) value;
+            return value;
         }
     }
 }
